reorganized views
moved them  into separate packages
