+++ src/main/java/trackour/trackour/views/admin/AdminErrorPage.java
@@ -1,0 +1,5 @@
+package trackour.trackour.views.admin;
+
+public class AdminErrorPage {
+    
+}
