+++ src/main/java/trackour/trackour/views/login/LoginPage.java
@@ -1,4 +1,4 @@
-package trackour.trackour.views.main;
+package trackour.trackour.views.login;
 
 import com.vaadin.flow.component.html.H1;
 import com.vaadin.flow.component.login.LoginForm;
